Update SYCL component for oneAPI 1.2 release (#429)
* Remove unused template names

* Simplify the DPC++ component

Simplify the description of the the DPC++ component.  We only need to
state that DPC++ is a combination of SYCL and a certain set of SYCL
extensions.  Remove the "test" column of the extension table since
most entries are just N/A.

* Rename from DPC++ to SYCL

Rename the "DPC++" component to "SYCL".  We no longer refer to "DPC++"
as a specification in its own right.  Rather, DPC++ is the name of an
implementation of the SYCL spec.

* Update SYCL extension table

This is the list of "supported" extensions in the DPC++ 2022.2 compiler
release.  Pointers to extension documents are now a tagged URL, so the
oneAPI spec refers to a specific copy of the extension spec even if the
version in the DPC++ repo evolves over time.